security: authorize notification actions and isolate trusted API helpers
